backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
PyPDF2
python-docx
requests
python-dotenv
transformers==4.40.0
torch>=2.0.0
google-generativeai
firebase-admin
sqlalchemy
psycopg2-binary
redis
celery
websockets
asyncio-mqtt
numpy
pandas
scikit-learn
nltk
spacy
python-jose[cryptography]
passlib[bcrypt]
